Dungeon crawler update (#530)
### Description

This PR improves the dungeon crawler tutorial, updates it to Godot 4.4,
updates certain areas to match current best practice, and fixes minor
correctness issues. Primarily:

- Updates project to Godot 4.4 compatibility
- Improves how weapon animation is handled, in a way that makes the
tutorial easier to follow by compacting a lot of steps onto a single
node
- Improves how UI is handled by introducing singleton nodes, which also
removes the previously bad way of accessing the UI node
- Fixes a bug in which enemies would only damage the player upon
entering the player hitbox, but would not continue to damage the player
if they remained in contact with it.
- Many minor correctness changes

Second version of this PR as the initial version was done without first
pulling from main... Which is also why this unfortunately has no commit
history.