fix: CONTRIBUTING.md in depth links failing to resolve (#580)
<!-- Thank you for opening a PR! We really appreciate you taking the
time to help out 🙌 -->

#### Description 

<!-- Please describe the change you are proposing, and why -->
I've updated the links within the contributing file to reflect what I
now believe to be the correct pages. The currently linked resources fail
to resolve and as such may disincentivize potential contributors who are
looking for further information.

I've also updated the wording from "Astro Docs", as it is unclear at
face value that this repository is what is being referred to. I feel
that the new wording more clearly states that the document is discussing
the repository it is housed in.